Move remaining code in `fusion._private` to `runtime._private`.
