+++ fusioncli/src/main/java/dev/ionfusion/fusioncli/cover/CoverageReportWriter.java
@@ -14,8 +14,7 @@ import com.amazon.ion.Span;
 import com.amazon.ion.SpanProvider;
 import com.amazon.ion.Timestamp;
 import com.amazon.ion.system.IonReaderBuilder;
-import dev.ionfusion.fusion._private.HtmlWriter;
-import dev.ionfusion.fusion._private.StreamWriter;
+import dev.ionfusion.runtime._private.io.StreamWriter;
 import dev.ionfusion.runtime.base.ModuleIdentity;
 import dev.ionfusion.runtime.base.SourceLocation;
 import java.io.IOException;
