+++ runtime/src/main/java/dev/ionfusion/runtime/_private/io/StreamWriter.java
@@ -1,7 +1,7 @@
 // Copyright Ion Fusion contributors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package dev.ionfusion.fusion._private;
+package dev.ionfusion.runtime._private.io;
 
 import static com.amazon.ion.system.IonTextWriterBuilder.UTF8;
