+++ runtime/src/main/java/dev/ionfusion/runtime/_private/util/InternMap.java
@@ -1,7 +1,7 @@
 // Copyright Ion Fusion contributors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package dev.ionfusion.fusion._private;
+package dev.ionfusion.runtime._private.util;
 
 import static java.util.Objects.requireNonNull;
