+++ runtime/src/test/java/dev/ionfusion/runtime/_private/util/InternedString.java
@@ -1,7 +1,7 @@
 // Copyright Ion Fusion contributors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package dev.ionfusion.fusion._private;
+package dev.ionfusion.runtime._private.util;
 
 import java.util.Objects;
